initial commit and scaffolding
